+++ src/main/java/Application.java
@@ -3,8 +3,8 @@ import java.util.Optional;
 import java.util.function.Supplier;
 import model.piece.Piece;
 import model.position.Position;
-import model.Team;
-import model.position.Score;
+import model.piece.Team;
+import model.piece.Score;
 import service.JanggiService;
 import view.InputView;
 import view.OutputView;
