+++ src/main/java/dao/PieceDao.java
@@ -4,15 +4,15 @@ import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.util.HashMap;
 import java.util.Map;
-import model.Team;
-import model.piece.Byeong;
-import model.piece.Cannon;
-import model.piece.Chariot;
-import model.piece.Elephant;
-import model.piece.General;
-import model.piece.Guard;
-import model.piece.Horse;
-import model.piece.Jol;
+import model.piece.Team;
+import model.piece.type.Byeong;
+import model.piece.type.Cannon;
+import model.piece.type.Chariot;
+import model.piece.type.Elephant;
+import model.piece.type.General;
+import model.piece.type.Guard;
+import model.piece.type.Horse;
+import model.piece.type.Jol;
 import model.piece.Piece;
 import model.position.Column;
 import model.position.Position;
