+++ src/main/java/model/navigator/JumpMoveNavigator.java
@@ -1,7 +1,7 @@
-package model.piece;
+package model.navigator;
 
 import java.util.List;
-import model.Movement;
+import model.position.Movement;
 import model.position.Position;
 
 public class JumpMoveNavigator {
