+++ src/main/java/model/piece/Piece.java
@@ -1,10 +1,7 @@
 package model.piece;
 
-import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
-import model.Movement;
-import model.Team;
 import model.position.Position;
 
 public abstract class Piece {
