+++ src/main/java/model/piece/Team.java
@@ -1,4 +1,4 @@
-package model;
+package model.piece;
 
 
 public enum Team {
