+++ src/main/java/model/piece/type/Guard.java
@@ -1,11 +1,13 @@
-package model.piece;
+package model.piece.type;
 
-import static model.Movement.*;
+import static model.position.Movement.*;
 
-import java.util.ArrayList;
 import java.util.List;
-import model.Movement;
-import model.Team;
+import model.position.Movement;
+import model.navigator.LimitedBasicMoveNavigator;
+import model.piece.Castle;
+import model.piece.Piece;
+import model.piece.Team;
 import model.position.Position;
 
 public class Guard extends Piece {
