+++ src/test/java/model/HorseTest.java
@@ -4,7 +4,8 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import java.util.List;
-import model.piece.Horse;
+import model.piece.type.Horse;
+import model.piece.Team;
 import model.position.Column;
 import model.position.Position;
 import model.position.Row;
